Moved setup pipeline classes and function from experiment.py to pipeline.py
