+++ Yank/pipeline.py
@@ -14,13 +14,18 @@ Utility functions to help setting up Yank configurations.
 #=============================================================================================
 
 import os
+import re
+import copy
 import inspect
 import logging
 import itertools
+import collections
 
 import numpy as np
 import openmmtools as mmtools
+import openmoltools as moltools
 from simtk import openmm, unit
+from simtk.openmm.app import PDBFile
 
 from . import utils
 
@@ -468,6 +473,948 @@ def read_system_files(positions_file_path, parameters_file_path, system_options,
         raise RuntimeError(err_msg)
 
     return system, parameters_file.topology, sampler_state
+
+
+# =============================================================================
+# SETUP PIPELINE UTILITY FUNCTIONS
+# =============================================================================
+
+# Map the OpenMM-style name for a solvent to the tleap
+# name compatible with the solvateBox command.
+_OPENMM_LEAP_SOLVENT_MODELS_MAP = {
+    'tip3p': 'TIP3PBOX',
+    'tip3pfb': 'TIP3PFBOX',
+    'tip4pew': 'TIP4PEWBOX',
+    'tip5p': 'TIP5PBOX',
+    'spce': 'SPCBOX',
+}
+
+
+def remove_overlap(mol_positions, *args, **kwargs):
+    """Remove any eventual overlap between a molecule and a set of others.
+
+    The method both randomly shifts and rotates the molecule (when overlapping atoms
+    are detected) until it does not clash with any other given molecule anymore. All
+    the others are kept fixed.
+
+    All the positions must be expressed in the same unit of measure.
+
+    Parameters
+    ----------
+    mol_positions : numpy.ndarray
+        An Nx3 array where, N is the number of atoms, containing the positions of
+        the atoms of the molecule that we want to not clash with the others.
+    min_distance : float
+        The minimum distance accepted to consider the molecule not clashing with
+        the others. Must be in the same unit of measure of the positions.
+    sigma : float
+        The maximum displacement for a single step. Must be in the same unit of
+        measure of the positions.
+
+    Other parameters
+    ----------------
+    args
+        A series of numpy.ndarrays containing the positions of the atoms of the
+        molecules that are kept fixed.
+
+    Returns
+    -------
+    x : numpy.ndarray
+        Positions of the atoms of the given molecules that do not clash.
+
+    """
+    x = np.copy(mol_positions)
+    sigma = kwargs.get('sigma', 1.0)
+    min_distance = kwargs.get('min_distance', 1.0)
+
+    # Try until we have a non-overlapping conformation w.r.t. all fixed molecules
+    while compute_min_dist(x, *args) <= min_distance:
+        # Compute center of geometry
+        x0 = x.mean(0)
+
+        # Randomize orientation of ligand.
+        Rq = mmtools.mcmc.MCRotationMove.generate_random_rotation_matrix()
+        x = ((Rq * np.matrix(x - x0).T).T + x0).A
+
+        # Choose a random displacement vector and translate
+        x += sigma * np.random.randn(3)
+
+    return x
+
+
+def pack_transformation(mol1_pos, mol2_pos, min_distance, max_distance):
+    """Compute an affine transformation that solve clashes and fit mol2 in the box.
+
+    The method randomly shifts and rotates mol2 until all its atoms are within
+    min_distance and max_distance from mol1. The position of mol1 is kept fixed.
+    Every 200 failed iterations, the algorithm increases max_distance by 50%. It
+    raise an exception after 1000 iterations.
+
+    All the positions must be expressed in the same unit of measure.
+
+    Parameters
+    ----------
+    mol1_pos : numpy.ndarray
+        An Nx3 array where, N is the number of atoms, containing the positions of
+        the atoms of the molecule that will be kept fixed.
+    mol2_pos : numpy.ndarray
+        An Nx3 array where, N is the number of atoms, containing the positions of
+        the atoms of the molecule that will be eventually moved.
+    min_distance : float
+        The minimum distance accepted to consider mol2 not clashing with mol1. It
+        must be in the same unit of measure of the positions.
+    max_distance : float
+        The maximum distance from mol1 to consider mol2 within the box. It must
+        be in the same unit of measure of the positions.
+
+    Returns
+    -------
+    transformation : numpy.ndarray
+        A 4x4 ndarray representing the affine transformation that translate and
+        rotate mol2.
+
+    """
+    translation = None  # we'll use this to check if we made changes to mol2_pos
+    transformation = np.identity(4)
+
+    # Compute center of geometry
+    x0 = mol2_pos.mean(0)
+
+    # Try until we have a non-overlapping conformation w.r.t. all fixed molecules
+    i = 0
+    min_dist, max_dist = compute_min_max_dist(mol1_pos, mol2_pos)
+    while min_dist < min_distance or max_distance <= max_dist:
+        # Select random atom of fixed molecule and use it to propose new x0 position
+        mol1_atom_idx = np.random.random_integers(0, len(mol1_pos) - 1)
+        translation = mol1_pos[mol1_atom_idx] + max_distance * np.random.randn(3) - x0
+
+        # Generate random rotation matrix
+        Rq = mmtools.mcmc.MCRotationMove.generate_random_rotation_matrix()
+
+        # Apply random transformation and test
+        x = ((Rq * np.matrix(mol2_pos - x0).T).T + x0).A + translation
+        min_dist, max_dist = compute_min_max_dist(mol1_pos, x)
+
+        # Check n iterations
+        i += 1
+        if i % 200 == 0:
+            max_distance *= 1.5
+        if i >= 1000:
+            err_msg = 'Cannot fit mol2 into solvation box!'
+            logger.error(err_msg)
+            raise RuntimeError(err_msg)
+
+    # Generate 4x4 affine transformation in molecule reference frame
+    if translation is not None:
+        transl_to_origin, transl_to_x0, rot_transl_matrix = (np.identity(4) for _ in range(3))
+        transl_to_origin[:3, 3] = -x0  # translate the molecule from x0 to origin
+        rot_transl_matrix[:3, :3] = Rq  # rotate molecule in origin
+        rot_transl_matrix[:3, 3] = translation  # translate molecule
+        transl_to_x0[:3, 3] = x0  # translate the molecule from origin to x0
+        transformation = transl_to_x0.dot(rot_transl_matrix.dot(transl_to_origin))
+
+    return transformation
+
+
+def pull_close(fixed_mol_pos, translated_mol_pos, min_bound, max_bound):
+    """Heuristic algorithm to quickly translate the ligand close to the receptor.
+
+    The distance of the ligand from the receptor here is defined as the shortest
+    Euclidean distance between an atom of the ligand and one of the receptor.
+    The molecules positions will not be modified if the ligand is already at a
+    distance in the interval [min_bound, max_bound].
+
+    Parameters
+    ----------
+    fixed_mol_pos : numpy.array
+        The positions of the molecule to keep fixed as a Nx3 array.
+    translated_mol_pos : numpy.array
+        The positions of the molecule to translate as a Nx3 array.
+    min_bound : float
+        Minimum distance from the receptor to the ligand. This should be high
+        enough for the ligand to not overlap the receptor atoms at the beginning
+        of the simulation.
+    max_bound : float
+        Maximum distance from the receptor to the ligand. This should be short
+        enough to make the ligand and the receptor interact since the beginning
+        of the simulation.
+
+    Returns
+    -------
+    translation : numpy.array
+        A 1x3 array containing the translation vector to apply to translated_mol_pos
+        to move the molecule at a distance between min_bound and max_bound from
+        fixed_mol_pos.
+
+    """
+
+    goal_distance = (min_bound + max_bound) / 2
+    trans_pos = copy.deepcopy(translated_mol_pos)  # positions that we can modify
+
+    # Find translation
+    final_translation = np.zeros(3)
+    while True:
+
+        # Compute squared distances between all atoms
+        # Each row is an array of distances from a translated atom to all fixed atoms
+        # We don't need to apply square root to everything
+        distances2 = np.array([((fixed_mol_pos - pos)**2).sum(1) for pos in trans_pos])
+
+        # Find closest atoms and their distance
+        min_idx = np.unravel_index(distances2.argmin(), distances2.shape)
+        min_dist = np.sqrt(distances2[min_idx])
+
+        # If closest atom is between boundaries translate ligand
+        if min_bound <= min_dist <= max_bound:
+            break
+
+        # Compute unit vector that connects receptor and ligand atom
+        if min_dist != 0:
+            direction = fixed_mol_pos[min_idx[1]] - trans_pos[min_idx[0]]
+        else:  # any deterministic direction
+            direction = np.array([1, 1, 1])
+        direction = direction / np.sqrt((direction**2).sum())  # normalize
+
+        if max_bound < min_dist:  # the atom is far away
+            translation = (min_dist - goal_distance) * direction
+            trans_pos += translation
+            final_translation += translation
+        elif min_dist < min_bound:  # the two molecules overlap
+            max_dist = np.sqrt(distances2.max())
+            translation = (max_dist + goal_distance) * direction
+            trans_pos += translation
+            final_translation += translation
+
+    return final_translation
+
+
+def strip_protons(input_file_path, output_file_path):
+    """Remove all hydrogens from PDB file and save the result."""
+    output_file = open(output_file_path, 'w')
+    with open(input_file_path, 'r') as input_file:
+        for line in input_file:
+            if not (line[:6] == 'ATOM  ' and (line[12] == 'H' or line[13] == 'H')):
+                output_file.write(line)
+    output_file.close()
+
+
+def read_csv_lines(file_path, lines):
+    """Return a list of CSV records.
+
+    The function takes care of ignoring comments and blank lines.
+
+    Parameters
+    ----------
+    file_path : str
+        The path to the CSV file.
+    lines : 'all' or int
+        The index of the line to read or 'all' to return
+        the list of all lines.
+
+    Returns
+    -------
+    records : str or list of str
+        The CSV record if lines is an integer, or a list of CSV
+        records if it is 'all'.
+
+    """
+    # Read all lines ignoring blank lines and comments.
+    with open(file_path, 'r') as f:
+        all_records = [line for line in f
+                       if bool(line) and not line.strip().startswith('#')]
+
+    if lines == 'all':
+        return all_records
+    return all_records[lines]
+
+
+# ==============================================================================
+# SETUP DATABASE
+# ==============================================================================
+
+class SetupDatabase:
+    """Provide utility functions to set up systems and molecules.
+
+    The object allows to access molecules, systems and solvents by 'id' and takes
+    care of parametrizing molecules and creating the AMBER prmtop and inpcrd files
+    describing systems.
+
+    Attributes
+    ----------
+    setup_dir : str
+        Path to the main setup directory. Changing this means changing the database.
+    molecules : dict
+        Dictionary of molecule_id -> molecule YAML description.
+    solvents : dict
+        Dictionary of solvent_id -> molecule YAML description.
+
+    """
+
+    SYSTEMS_DIR = 'systems'
+    MOLECULES_DIR = 'molecules'
+    CLASH_THRESHOLD = 1.5  # distance in Angstroms to consider two atoms clashing
+
+    def __init__(self, setup_dir, molecules=None, solvents=None, systems=None):
+        """Initialize the database.
+
+        Parameters
+        ----------
+        setup_dir : str
+            Path to the main setup directory.
+        molecules : dict
+            YAML description of the molecules (default is None).
+        solvents : dict
+            YAML description of the solvents (default is None).
+
+        """
+        self.setup_dir = setup_dir
+        self.molecules = molecules
+        self.solvents = solvents
+        self.systems = systems
+
+        # Private attributes
+        self._pos_cache = {}  # cache positions of molecules
+        self._processed_mols = set()  # keep track of parametrized molecules
+
+    def get_molecule_dir(self, molecule_id):
+        """Return the directory where the parameter files are stored.
+
+        Parameters
+        ----------
+        molecule_id : str
+            The ID of the molecule.
+
+        Returns
+        -------
+        str
+            The path to the molecule directory.
+        """
+        return os.path.join(self.setup_dir, self.MOLECULES_DIR, molecule_id)
+
+    def get_system_files_paths(self, system_id):
+        """Return the paths to the systems files.
+
+        Parameters
+        ----------
+        system_id : str
+            The ID of the system.
+
+        Returns
+        -------
+        system_files_paths : list of namedtuple
+            Elements of the list contain the paths to the system files for
+            each phase. Each namedtuple contains the fields position_path (e.g.
+            inpcrd, gro, or pdb) and parameters_path (e.g. prmtop, top, or xml).
+
+        """
+        Paths = collections.namedtuple('Paths', ['position_path', 'parameters_path'])
+        system_dir = os.path.join(self.setup_dir, self.SYSTEMS_DIR, system_id)
+        if 'receptor' in self.systems[system_id]:
+            system_files_paths = [
+                Paths(position_path=os.path.join(system_dir, 'complex.inpcrd'),
+                      parameters_path=os.path.join(system_dir, 'complex.prmtop')),
+                Paths(position_path=os.path.join(system_dir, 'solvent.inpcrd'),
+                      parameters_path=os.path.join(system_dir, 'solvent.prmtop'))
+            ]
+        elif 'solute' in self.systems[system_id]:
+            system_files_paths = [
+                Paths(position_path=os.path.join(system_dir, 'solvent1.inpcrd'),
+                      parameters_path=os.path.join(system_dir, 'solvent1.prmtop')),
+                Paths(position_path=os.path.join(system_dir, 'solvent2.inpcrd'),
+                      parameters_path=os.path.join(system_dir, 'solvent2.prmtop'))
+            ]
+        else:
+            system_files_paths = [
+                Paths(position_path=self.systems[system_id]['phase1_path'][0],
+                      parameters_path=self.systems[system_id]['phase1_path'][1]),
+                Paths(position_path=self.systems[system_id]['phase2_path'][0],
+                      parameters_path=self.systems[system_id]['phase2_path'][1])
+            ]
+
+        return system_files_paths
+
+    def is_molecule_setup(self, molecule_id):
+        """Check whether the molecule has been processed previously.
+
+        The molecule must be set up if it needs to be parametrize by antechamber
+        (and the gaff.mol2 and frcmod files do not exist), if the molecule must be
+        generated by OpenEye, or if it needs to be extracted by a multi-molecule file.
+
+        An example to clarify the difference between the two return values: a protein
+        in a single-frame pdb does not have to be processed (since it does not go through
+        antechamber) thus the function will return is_setup=True and is_processed=False.
+
+        Parameters
+        ----------
+        molecule_id : str
+            The id of the molecule.
+
+        Returns
+        -------
+        is_setup : bool
+            True if the molecule's parameter files have been specified by the user
+            or if they have been generated by SetupDatabase.
+        is_processed : bool
+            True if parameter files have been generated previously by SetupDatabase
+            (i.e. if the parameter files were not manually specified by the user).
+
+        """
+
+        # The only way to check if we processed the molecule in the current run is
+        # through self._processed_mols as 'parameters' will be changed after setup
+        if molecule_id in self._processed_mols:
+            return True, True
+
+        # Some convenience variables
+        molecule_descr = self.molecules[molecule_id]
+        molecule_dir = self.get_molecule_dir(molecule_id)
+        molecule_id_path = os.path.join(molecule_dir, molecule_id)
+        try:
+            extension = os.path.splitext(molecule_descr['filepath'])[1]
+        except KeyError:
+            extension = None
+
+        # The following checks must be performed in reverse order w.r.t. how they
+        # are executed in _setup_molecules()
+
+        files_to_check = {}
+
+        # If the molecule must go through antechamber we search for its output
+        if 'antechamber' in molecule_descr:
+            files_to_check = [('filepath', molecule_id_path + '.gaff.mol2'),
+                              (['leap', 'parameters'], molecule_id_path + '.frcmod')]
+
+        # If the molecule must be generated by OpenEye, a mol2 should have been created
+        elif extension is None or extension == '.smiles' or extension == '.csv':
+            files_to_check = [('filepath', molecule_id_path + '.mol2')]
+
+        # If we have to strip the protons off a PDB, a new PDB should have been created
+        elif 'strip_protons' in molecule_descr and molecule_descr['strip_protons']:
+            files_to_check = [('filepath', molecule_id_path + '.pdb')]
+
+        # If a single structure must be extracted we search for output
+        elif 'select' in molecule_descr:
+            files_to_check = [('filepath', molecule_id_path + extension)]
+
+        # Check if this needed to be processed at all
+        if not files_to_check:
+            return True, False
+
+        # Check if all output files exist
+        all_file_exist = True
+        for descr_key, file_path in files_to_check:
+            all_file_exist &= os.path.isfile(file_path) and os.path.getsize(file_path) > 0
+            if all_file_exist:  # Make sure internal description is correct
+                try:
+                    molecule_descr[descr_key] = file_path
+                except TypeError:  # nested key, list are unhashable
+                    molecule_descr[descr_key[0]][descr_key[1]].append(file_path)
+
+        # Compute and update small molecule net charge
+        if all_file_exist:
+            extension = os.path.splitext(molecule_descr['filepath'])[1]
+            # TODO what if this is a peptide? This should be computed in get_system()
+            if extension == '.mol2':
+                molecule_descr['net_charge'] = utils.Mol2File(molecule_descr['filepath']).net_charge
+
+        return all_file_exist, all_file_exist
+
+    def is_system_setup(self, system_id):
+        """Check whether the system has been already processed.
+
+        Parameters
+        ----------
+        system_id : str
+            The ID of the system.
+
+        Returns
+        -------
+        is_setup : bool
+            True if the system is ready to be used for an experiment. Either because
+            the system has directly provided the system files, or because it already
+            went through the setup pipeline.
+        is_processed : bool
+            True if the system has already gone through the setup pipeline.
+
+        """
+        if 'ligand' in self.systems[system_id] or 'solute' in self.systems[system_id]:
+            system_files_paths = self.get_system_files_paths(system_id)
+            is_setup = (os.path.exists(system_files_paths[0].position_path) and
+                        os.path.exists(system_files_paths[0].parameters_path) and
+                        os.path.exists(system_files_paths[1].position_path) and
+                        os.path.exists(system_files_paths[1].parameters_path))
+            return is_setup, is_setup
+        else:
+            return True, False
+
+    def get_system(self, system_id):
+        """Make sure that the system files are set up and return the system folder.
+
+        If necessary, create the prmtop and inpcrd files from the given components.
+        The system files are generated with tleap. If no molecule specifies a general
+        force field, leaprc.ff14SB is loaded.
+
+        Parameters
+        ----------
+        system_id : str
+            The ID of the system.
+
+        Returns
+        -------
+        system_files_paths : list of namedtuple
+            Elements of the list contain the paths to the system files for
+            each phase. Each namedtuple contains the fields position_path (e.g.
+            inpcrd, gro, or pdb) and parameters_path (e.g. prmtop, top, or xml).
+
+        """
+        # Check if system has been already processed
+        system_files_paths = self.get_system_files_paths(system_id)
+        if self.is_system_setup(system_id)[0]:
+            return system_files_paths
+
+        system_descr = self.systems[system_id]
+
+        if 'receptor' in system_descr:  # binding free energy calculation
+            receptor_id = system_descr['receptor']
+            ligand_id = system_descr['ligand']
+            solvent_id = system_descr['solvent']
+            system_parameters = system_descr['leap']['parameters']
+
+            # solvent phase
+            self._setup_system(system_files_paths[1].position_path, False,
+                               0, system_parameters, solvent_id, ligand_id)
+
+            try:
+                alchemical_charge = int(round(self.molecules[ligand_id]['net_charge']))
+            except KeyError:
+                alchemical_charge = 0
+
+            # complex phase
+            self._setup_system(system_files_paths[0].position_path,
+                               system_descr['pack'], alchemical_charge,
+                               system_parameters,  solvent_id, receptor_id,
+                               ligand_id)
+        else:  # partition/solvation free energy calculation
+            solute_id = system_descr['solute']
+            solvent1_id = system_descr['solvent1']
+            solvent2_id = system_descr['solvent2']
+            system_parameters = system_descr['leap']['parameters']
+
+            # solvent1 phase
+            self._setup_system(system_files_paths[0].position_path, False,
+                               0, system_parameters, solvent1_id, solute_id)
+
+            # solvent2 phase
+            self._setup_system(system_files_paths[1].position_path, False,
+                               0, system_parameters, solvent2_id, solute_id)
+
+        return system_files_paths
+
+    def _generate_molecule(self, molecule_id):
+        """Generate molecule using the OpenEye toolkit from name or smiles.
+
+        The molecule is charged with OpenEye's recommended AM1BCC charge
+        selection scheme and it is saved into the OpenEye molecules cache.
+
+        Parameters
+        ----------
+        molecule_id : str
+            The id of the molecule as given in the YAML script
+
+        Returns
+        -------
+        molecule : OEMol
+            The generated molecule.
+
+        """
+        mol_descr = self.molecules[molecule_id]  # molecule description
+        try:
+            if 'name' in mol_descr:
+                molecule = moltools.openeye.iupac_to_oemol(mol_descr['name'])
+            elif 'smiles' in mol_descr:
+                molecule = moltools.openeye.smiles_to_oemol(mol_descr['smiles'])
+            molecule = moltools.openeye.get_charges(molecule, keep_confs=1)
+        except ImportError as e:
+            error_msg = ('requested molecule generation from name or smiles but '
+                         'could not find OpenEye toolkit: ' + str(e))
+            logger.error(error_msg)
+            raise RuntimeError(error_msg)
+
+        return molecule
+
+    def _setup_molecules(self, *args):
+        """Set up the files needed to generate the system for all the molecules.
+
+        If OpenEye tools are installed, this generate the molecules when the source is
+        not a file. If two (or more) molecules generated by OpenEye have overlapping
+        atoms, the molecules are randomly shifted and rotated until the clash is resolved.
+        With the OpenEye toolkit installed, we also perform a sanity check to verify that
+        the molecules from files do not have overlapping atoms. An Exception is raised if
+        this is not the case.
+
+        If the Schrodinger's suite is install, this can enumerate tautomeric and protonation
+        states with epik when requested.
+
+        This also parametrize the molecule with antechamber when requested.
+
+        Other parameters
+        ----------------
+        args
+            All the molecules ids that compose the system. These molecules are the only
+            ones considered when trying to resolve the overlapping atoms.
+
+        """
+
+        for mol_id in args:
+            net_charge = None  # used by antechamber
+            mol_descr = self.molecules[mol_id]
+
+            # Have we already processed this molecule? Do we have to do it at all?
+            # We don't want to create the output folder if we don't need to
+            if self.is_molecule_setup(mol_id)[0]:
+                continue
+
+            # Create output directory if it doesn't exist
+            mol_dir = self.get_molecule_dir(mol_id)
+            if not os.path.exists(mol_dir):
+                os.makedirs(mol_dir)
+
+            try:
+                extension = os.path.splitext(mol_descr['filepath'])[1]
+            except KeyError:
+                extension = None
+
+            # Extract single model if this is a multi-model file
+            if 'select' in mol_descr:
+                model_idx = mol_descr['select']
+                single_file_path = os.path.join(mol_dir, mol_id + extension)
+                if extension == '.pdb':
+                    # Create single-model PDB file
+                    pdb_file = PDBFile(mol_descr['filepath'])
+                    with open(single_file_path, 'w') as f:
+                        PDBFile.writeHeader(pdb_file.topology, file=f)
+                        PDBFile.writeModel(pdb_file.topology, pdb_file.getPositions(frame=model_idx), file=f)
+                    # We might as well already cache the positions
+                    self._pos_cache[mol_id] = pdb_file.getPositions(asNumpy=True, frame=model_idx) / unit.angstrom
+
+                elif extension == '.smiles' or extension == '.csv':
+                    # Extract the correct line and save it in a new file.
+                    smiles_line = read_csv_lines(mol_descr['filepath'], lines=model_idx)
+                    with open(single_file_path, 'w') as f:
+                        f.write(smiles_line)
+
+                elif extension == '.mol2' or extension == '.sdf':
+                    if not utils.is_openeye_installed(oetools=('oechem',)):
+                        raise RuntimeError('Cannot support {} files selection without OpenEye'.format(
+                                extension[1:]))
+                    oe_molecule = utils.read_oe_molecule(mol_descr['filepath'], conformer_idx=model_idx)
+                    if extension == '.mol2':
+                        mol_names = utils.Mol2File(mol_descr['filepath']).resnames
+                        utils.write_oe_molecule(oe_molecule, single_file_path, mol2_resname=mol_names[model_idx])
+                    else:
+                        utils.write_oe_molecule(oe_molecule, single_file_path)
+
+                else:
+                    raise RuntimeError('Model selection is not supported for {} files'.format(extension[1:]))
+
+                # Save new file path
+                mol_descr['filepath'] = single_file_path
+
+            # Strip off protons if required
+            if 'strip_protons' in mol_descr and mol_descr['strip_protons']:
+                if extension != '.pdb':
+                    raise RuntimeError('Cannot strip protons off {} files.'.format(extension[1:]))
+                output_file_path = os.path.join(mol_dir, mol_id + '.pdb')
+                strip_protons(mol_descr['filepath'], output_file_path)
+                mol_descr['filepath'] = output_file_path
+
+            # Generate missing molecules with OpenEye. At the end of parametrization
+            # we update the 'filepath' key also for OpenEye-generated molecules so
+            # we don't need to keep track of the molecules we have already generated
+            if extension is None or extension == '.smiles' or extension == '.csv':
+                if not utils.is_openeye_installed(oetools=('oechem', 'oeiupac', 'oequacpac', 'oeomega')):
+                    if extension is None:
+                        raise RuntimeError('Cannot generate molecule {} without OpenEye licensed with '
+                                           'OEChem, OEIUPAC, OEOmega, and OEQuacPack.'.format(mol_id))
+                    else:
+                        raise RuntimeError('Cannot support {} files without OpenEye licensed with '
+                                           'OEChem, OEIUPAC, OEOmega, and OEQuacPack.'.format(extension[1:]))
+
+                # Retrieve the first SMILES string (eventually extracted
+                # while handling of the 'select' keyword above)
+                if extension is not None:
+                    # Get first record in CSV file.
+                    first_line = read_csv_lines(mol_descr['filepath'], lines=0)
+
+                    # Automatically detect if delimiter is comma or semicolon
+                    for delimiter in ',;':
+                        logger.debug("Attempt to parse smiles file with delimiter '{}'".format(delimiter))
+                        line_fields = first_line.split(delimiter)
+                        # If there is only one column, take that, otherwise take second
+                        if len(line_fields) > 1:
+                            smiles_str = line_fields[1].strip()
+                        else:
+                            smiles_str = line_fields[0].strip()
+
+                        # try to generate the smiles and try new delimiter if it fails
+                        mol_descr['smiles'] = smiles_str
+                        try:
+                            oe_molecule = self._generate_molecule(mol_id)
+                            break
+                        except (ValueError, RuntimeError):
+                            oe_molecule = None
+
+                    # Raise an error if no delimiter worked
+                    if oe_molecule is None:
+                        raise RuntimeError('Cannot detect SMILES file format.')
+                else:
+                    # Generate molecule from mol_descr['smiles']
+                    oe_molecule = self._generate_molecule(mol_id)
+
+                # Cache atom positions
+                self._pos_cache[mol_id] = utils.get_oe_mol_positions(oe_molecule)
+
+                # Write OpenEye generated molecules in mol2 files
+                # We update the 'filepath' key in the molecule description
+                mol_descr['filepath'] = os.path.join(mol_dir, mol_id + '.mol2')
+
+                # We set the residue name as the first three uppercase letters of mol_id
+                residue_name = re.sub('[^A-Za-z]+', '', mol_id.upper())[:3]
+                moltools.openeye.molecule_to_mol2(oe_molecule, mol_descr['filepath'],
+                                                  residue_name=residue_name)
+
+            # Enumerate protonation states with epik
+            if 'epik' in mol_descr:
+                epik_base_path = os.path.join(mol_dir, mol_id + '-epik.')
+                epik_mae_file = epik_base_path + 'mae'
+                epik_mol2_file = epik_base_path + 'mol2'
+                epik_sdf_file = epik_base_path + 'sdf'
+
+                # Run epik and convert from maestro to both mol2 and sdf
+                # to not lose neither the penalties nor the residue name
+                epik_kwargs = mol_descr['epik']
+                moltools.schrodinger.run_epik(mol_descr['filepath'], epik_mae_file, **epik_kwargs)
+                moltools.schrodinger.run_structconvert(epik_mae_file, epik_sdf_file)
+                moltools.schrodinger.run_structconvert(epik_mae_file, epik_mol2_file)
+
+                # Save new net charge from the i_epik_Tot_Q property
+                net_charge = int(moltools.schrodinger.run_proplister(epik_sdf_file)[0]['i_epik_Tot_Q'])
+
+                # Keep filepath consistent
+                mol_descr['filepath'] = epik_mol2_file
+
+            # Antechamber does not support sdf files so we need to convert them
+            extension = os.path.splitext(mol_descr['filepath'])[1]
+            if extension == '.sdf':
+                if not utils.is_openeye_installed(oetools=('oechem',)):
+                    raise RuntimeError('Cannot support sdf files without OpenEye OEChem')
+                mol2_file_path = os.path.join(mol_dir, mol_id + '.mol2')
+                oe_molecule = utils.read_oe_molecule(mol_descr['filepath'])
+
+                # We set the residue name as the first three uppercase letters of mol_id
+                residue_name = re.sub('[^A-Za-z]+', '', mol_id.upper())[:3]
+                moltools.openeye.molecule_to_mol2(oe_molecule, mol2_file_path,
+                                                  residue_name=residue_name)
+
+                # Update filepath information
+                mol_descr['filepath'] = mol2_file_path
+
+            # Parametrize the molecule with antechamber
+            if 'antechamber' in mol_descr:
+                # Generate charges with OpenEye if requested
+                if 'openeye' in mol_descr:
+                    if not utils.is_openeye_installed(oetools=('oechem', 'oequacpac', 'oeomega')):
+                        err_msg = ('Cannot find OpenEye toolkit with OEChem and OEQuacPac to compute charges '
+                                   'for molecule {}').format(mol_id)
+                        logger.error(err_msg)
+                        raise RuntimeError(err_msg)
+                    mol2_file_path = os.path.join(mol_dir, mol_id + '.mol2')
+                    oe_molecule = utils.read_oe_molecule(mol_descr['filepath'])
+
+                    # Setting keep_confs = None keeps the original conformation
+                    oe_molecule = moltools.openeye.get_charges(oe_molecule, keep_confs=None)
+                    residue_name = utils.Mol2File(mol_descr['filepath']).resname
+                    moltools.openeye.molecule_to_mol2(oe_molecule, mol2_file_path,
+                                                      residue_name=residue_name)
+
+                    utils.Mol2File(mol2_file_path).net_charge = None  # normalize charges
+                    net_charge = None  # we don't need Epik's net charge
+                    mol_descr['filepath'] = mol2_file_path
+
+                # Generate parameters
+                charge_method = mol_descr['antechamber']['charge_method']
+                input_mol_path = os.path.abspath(mol_descr['filepath'])
+                with moltools.utils.temporary_cd(mol_dir):
+                    moltools.amber.run_antechamber(mol_id, input_mol_path,
+                                                   charge_method=charge_method,
+                                                   net_charge=net_charge)
+
+                # Save new parameters paths
+                mol_descr['filepath'] = os.path.join(mol_dir, mol_id + '.gaff.mol2')
+                mol_descr['leap']['parameters'].append(os.path.join(mol_dir, mol_id + '.frcmod'))
+
+                # Normalize charges if not done before
+                if 'openeye' not in mol_descr:
+                    utils.Mol2File(mol_descr['filepath']).net_charge = None
+
+            # Determine small molecule net charge
+            extension = os.path.splitext(mol_descr['filepath'])[1]
+            if extension == '.mol2':
+                # TODO what if this is a peptide? this should be computed in get_system()
+                mol_descr['net_charge'] = utils.Mol2File(mol_descr['filepath']).net_charge
+
+            # Keep track of processed molecule
+            self._processed_mols.add(mol_id)
+
+    def _setup_system(self, system_file_path, pack, alchemical_charge,
+                      system_parameters, solvent_id, *molecule_ids):
+        """Setup a system and create its prmtop/inpcrd files.
+
+        IMPORTANT: This function does not check if it's about to overwrite
+        files. Use get_system() for safe setup.
+
+        Parameters
+        ----------
+        system_file_path : str
+            The path to either the prmtop or inpcrd output file. The other one
+            will be saved in the same folder with the same base name.
+        pack : bool
+            True to automatically solve atom clashes and reduce box dimension.
+        alchemical_charge : int
+            Number of counterions to alchemically modify during the simulation.
+        system_parameters : list of str
+            Contain the parameters file that must be loaded in tleap for the
+            system in addition to the molecule-specific ones.
+        solvent_id : str
+            The ID of the solvent.
+        *molecule_ids : list-like of str
+            List the IDs of the molecules to pack together in the system.
+
+        """
+        solvent = self.solvents[solvent_id]
+        self._setup_molecules(*molecule_ids)  # Be sure molecules are set up
+
+        # Create tleap script
+        tleap = utils.TLeap()
+
+        # Load all parameters
+        # --------------------
+        tleap.new_section('Load parameters')
+
+        for mol_id in molecule_ids:
+            molecule_parameters = self.molecules[mol_id]['leap']['parameters']
+            tleap.load_parameters(*molecule_parameters)
+
+        tleap.load_parameters(*system_parameters)
+
+        # Load molecules and create complexes
+        # ------------------------------------
+        tleap.new_section('Load molecules')
+        for mol_id in molecule_ids:
+            tleap.load_unit(name=mol_id, file_path=self.molecules[mol_id]['filepath'])
+
+        if len(molecule_ids) > 1:
+            # Check that molecules don't have clashing atoms. Also, if the ligand
+            # is too far away from the molecule we want to pull it closer
+            # TODO this check should be available even without OpenEye
+            if pack and utils.is_openeye_installed(oetools=('oechem',)):
+
+                # Load atom positions of all molecules
+                positions = [0 for _ in molecule_ids]
+                for i, mol_id in enumerate(molecule_ids):
+                    if mol_id not in self._pos_cache:
+                        self._pos_cache[mol_id] = utils.get_oe_mol_positions(
+                                utils.read_oe_molecule(self.molecules[mol_id]['filepath']))
+                    positions[i] = self._pos_cache[mol_id]
+
+                # Find and apply the transformation to fix clashing
+                # TODO this doesn't work with more than 2 molecule_ids
+                try:
+                    max_dist = solvent['clearance'].value_in_unit(unit.angstrom) / 1.5
+                except KeyError:
+                    max_dist = 10.0
+                transformation = pack_transformation(positions[0], positions[1],
+                                                     self.CLASH_THRESHOLD, max_dist)
+                if (transformation != np.identity(4)).any():
+                    logger.warning('Changing starting positions for {}.'.format(molecule_ids[1]))
+                    tleap.new_section('Fix clashing atoms')
+                    tleap.transform(molecule_ids[1], transformation)
+
+            # Create complex
+            tleap.new_section('Create complex')
+            tleap.combine('complex', *molecule_ids)
+            unit_to_solvate = 'complex'
+        else:
+            unit_to_solvate = molecule_ids[0]
+
+        # Configure solvent
+        # ------------------
+        if solvent['nonbonded_method'] == openmm.app.NoCutoff:
+            try:
+                implicit_solvent = solvent['implicit_solvent']
+            except KeyError:  # vacuum
+                pass
+            else:  # implicit solvent
+                tleap.new_section('Set GB radii to recommended values for OBC')
+                tleap.add_commands('set default PBRadii {}'.format(
+                    get_leap_recommended_pbradii(implicit_solvent)))
+        else:  # explicit solvent
+            tleap.new_section('Solvate systems')
+
+            # Add alchemically modified ions
+            if alchemical_charge != 0:
+                try:
+                    if alchemical_charge > 0:
+                        ion = solvent['negative_ion']
+                    else:
+                        ion = solvent['positive_ion']
+                except KeyError:
+                    err_msg = ('Found charged ligand but no indications for ions in '
+                               'solvent {}').format(solvent_id)
+                    logger.error(err_msg)
+                    raise RuntimeError(err_msg)
+                tleap.add_ions(unit=unit_to_solvate, ion=ion, num_ions=abs(alchemical_charge))
+
+            # Neutralizing solvation box
+            if 'positive_ion' in solvent:
+                tleap.add_ions(unit=unit_to_solvate, ion=solvent['positive_ion'])
+            if 'negative_ion' in solvent:
+                tleap.add_ions(unit=unit_to_solvate, ion=solvent['negative_ion'])
+
+            # Solvate unit. Solvent models different than tip3p need parameter modifications.
+            solvent_model = solvent['solvent_model']
+            if not solvent_model.startswith('tip3p'):
+                tleap.load_parameters('frcmod.' + solvent_model)
+            leap_solvent_model = _OPENMM_LEAP_SOLVENT_MODELS_MAP[solvent_model]
+            clearance = float(solvent['clearance'].value_in_unit(unit.angstroms))
+            tleap.solvate(unit=unit_to_solvate, solvent_model=leap_solvent_model, clearance=clearance)
+
+        # Check charge
+        tleap.new_section('Check charge')
+        tleap.add_commands('check ' + unit_to_solvate)
+
+        # Save output files
+        # ------------------
+        system_dir = os.path.dirname(system_file_path)
+        base_file_path = os.path.basename(system_file_path).split('.')[0]
+        base_file_path = os.path.join(system_dir, base_file_path)
+
+        # Create output directory
+        if not os.path.exists(system_dir):
+            os.makedirs(system_dir)
+
+        # Save prmtop, inpcrd and reference pdb files
+        tleap.new_section('Save prmtop and inpcrd files')
+        tleap.save_unit(unit_to_solvate, system_file_path)
+        tleap.save_unit(unit_to_solvate, base_file_path + '.pdb')
+
+        # Save tleap script for reference
+        tleap.export_script(base_file_path + '.leap.in')
+
+        # Run tleap and log warnings
+        warnings = tleap.run()
+        for warning in warnings:
+            logger.warning('TLeap: ' + warning)
 
 
 if __name__ == '__main__':
